refactor login component with thunk moving fetch to actions
